requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-telegram-bot==22.5
playwright==1.49.1
httpx==0.28.1
beautifulsoup4==4.12.3
